Removed RolePermission and Added RoomUserRole
