--- src/AppBundle/Repository/RolePermissionRepository.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace AppBundle\Repository;
-
-/**
- * RolePermissionRepository
- *
- * This class was generated by the Doctrine ORM. Add your own custom
- * repository methods below.
- */
-class RolePermissionRepository extends \Doctrine\ORM\EntityRepository
-{
-}
